+++ src/input_parsers/TensorUtils.h
@@ -97,7 +97,7 @@ TensorShape getMultidirectionalBroadcastShape( TensorShape shape1, TensorShape s
 
 TensorIndices broadcastIndex ( TensorShape currentShape, TensorShape broadcastShape, TensorIndices broadcastIndices );
 
-TensorIndex unsignIndex ( TensorShape shape, SignedTensorIndex signedIndex );
+TensorIndex unsignIndex ( unsigned int size, SignedTensorIndex signedIndex );
 
 Permutation reversePermutation( unsigned int size );